+++ scripts/macros/spells/conjureElemental.js
@@ -14,7 +14,7 @@ export async function conjureElemental({speaker, actor, token, character, item, 
     let userId = game.settings.get('chris-premades', 'LastGM');
     if (game.settings.get('chris-premades', 'Player Chooses Conjures')) userId = game.userId;
     if (!userId) return;
-    let sourceActors = await chris.remoteDocumentDialog(userId, 'Select Summon', actors);
+    let sourceActors = await chris.remoteDocumentDialog(userId, 'Select Summon', actors, false, true, true);
     if (!sourceActors) return;
     let updates = {
         'token': {